🌱 Split CRDDrillDown.tsx (695 → 236 lines)
Extract presentational sub-components into CRDDrillDown.parts.tsx, shared
types and getConditionStyle into CRDDrillDown.types.ts, and all remote data
loading (versions, conditions, instances, schema) into useCRDDrillDown.ts.

Pure extraction — drill-down registration and useDrillDown() props unchanged.
Hook calls in the view drop from 13 to 6.
